Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    int i;
-    scanf("%d",&i);
+    float i;
+    scanf("%f",&i);
     printf("%f",i);
 
     return 0;
